+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CG.Blazor.Forms
 ---
 
+## 2022.1
+
+* I switched the runtime to .net 6.0
+* bug fixes
+
 ## 2021.2
 
 * I added support for generating vanilla HTML elements.